+++ classical_preprocessing/imaging_2d/mermed_encoder.py
@@ -185,7 +185,7 @@ class MerMEDEncoder(MedicalImageEncoder):
         with torch.no_grad():
             embeddings_tensor = self.model(image_batch)
 
-        embeddings_arr = embeddings_tensor.cpu().numpy().astype(np.float64)
+        embeddings_arr = embeddings_tensor.cpu().numpy().astype(np.float32)
 
         if not np.isfinite(embeddings_arr).all():
             raise ValueError("MerMED image encoder output contains NaN or Inf values.")
