+++ classical_preprocessing/imaging_2d/pipeline.py
@@ -132,7 +132,7 @@ class Imaging2DPipeline:
             sub_embeddings = self.encoder.encode(batch_tensor)
             all_embeddings.append(sub_embeddings)
 
-        embeddings_matrix = np.vstack(all_embeddings).astype(np.float64)
+        embeddings_matrix = np.vstack(all_embeddings).astype(np.float32)
 
         if not np.isfinite(embeddings_matrix).all():
             raise ValueError("Generated 2D image embeddings contain NaN or Inf values.")
